turn partition into directory module
